Remove the Ostree local remote
To build the container preinstalled image we used 4 ostree repo:
 * a local repo
 * a remote local repo
 * a remote repo
 * the image repo

What we did to create a new version of the container (or the OS):
 * we pull the remote local repo from the local repo
 * we commit the new version on the local repo
 * we push this new version on the remote local repo

Then to create the container preinstalled image:
 * we pull the remote local repo from the image repo
 * we remove that distant to add the remote repo as the distant

It is complicated and the config.cfg has to manage the two different remote (same
port at least). We remove the remote local repo to use the real remote repo. For
the Fullmetalupdate demo it still can be on the developper computer.